% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{first_noncon_analysis}
\alias{first_noncon_analysis}
\title{Perform First Non-Conformity Analysis on Sample Data}
\usage{
first_noncon_analysis(df, col_name, nominal_quantity, batch_size)
}
\arguments{
\item{df}{A data frame containing the sample data for analysis.}

\item{col_name}{A string representing the name of the column to analyze within the data frame.}

\item{nominal_quantity}{A numeric value indicating the nominal quantity of the product being produced.}

\item{batch_size}{A numeric value indicating the size of the batch being produced.}
}
\value{
A list with two elements: the decision on whether the batch should be accepted, rejected,
or subjected to further analysis, and the input data frame with a new column indicating whether each
value in the column being analyzed is acceptable, non-conforming, or unacceptable.
}
\description{
Given a data frame of sample data, a column name, a nominal quantity, and a batch size,
calculates the number of non-conformities and determines whether the batch should be
accepted, rejected, or subjected to further analysis.
}
\examples{
df <- data.frame(value = c(rep(100, 25),c(99, 99, 99, 99, 91)))
result <- first_noncon_analysis(df, "value", 100, 100)
result$decision
result$df
}
