% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{get_limits}
\alias{get_limits}
\title{Get Acceptance and Rejection Limits for Batch Size and Analysis Type}
\usage{
get_limits(batch_size, analysis)
}
\arguments{
\item{batch_size}{A numeric value indicating the size of the batch for which
the acceptance and rejection limits are to be returned.}

\item{analysis}{A string indicating the analysis type for which the
acceptance and rejection limits are to be returned. Must be one of "first" or
"second".}
}
\value{
A list with two elements, named "analysis_acceptance" and "analysis_rejection",
representing the acceptance and rejection limits for the specified batch size
and analysis type.
If the batch size or analysis type is outside the range of valid values, an error
will be thrown with a message indicating the nature of the problem.
}
\description{
Given a batch size and an analysis type (first or second), returns a list of
the acceptance and rejection limits for that batch size and analysis type.
}
\examples{
get_limits(1500, "first")
get_limits(100, "second")
}
