% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{get_mean_analysis_plot}
\alias{get_mean_analysis_plot}
\title{Create an confidence interval chart}
\usage{
get_mean_analysis_plot(
  batch_size,
  nominal_quantity,
  mean_val,
  std_dev,
  computed_student_dist
)
}
\arguments{
\item{batch_size}{Number representing the batch size}

\item{nominal_quantity}{Number representing the target nominal quantity}

\item{mean_val}{Number representing the estimated sample mean}

\item{std_dev}{Number representing the estimated sample standard deviation}

\item{computed_student_dist}{Number representing the computed student dist limit}
}
\value{
An interactive control chart using plotly
}
\description{
This function uses sample data, a value column, and a categorical column to create an interactive chart.
}
\examples{
get_mean_analysis_plot(100, 30, 30, 1, 0.503)

}
