% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{get_sample_sizes}
\alias{get_sample_sizes}
\title{Get Sample Sizes}
\usage{
get_sample_sizes(batch_size, analysis)
}
\arguments{
\item{batch_size}{A numeric value indicating the batch size for which the sample size
is to be calculated.}

\item{analysis}{A string value indicating the type of analysis for which the sample size
is to be calculated. Must be either "first" or "second".}
}
\value{
A numeric value indicating the required sample size for the given batch size
and analysis type. If the batch size is outside the range of the reference table, or if
the analysis parameter is not valid, the function returns an error message.
}
\description{
Calculates the required sample size for a given batch size and analysis type, based
on a reference table of sample sizes.
}
\examples{
get_sample_sizes(300, "first")
get_sample_sizes(1500, "second")
}
