% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{max_errors}
\alias{max_errors}
\title{Maximum error tolerances for nominal quantities.}
\format{
A data frame with 7 rows and 4 variables:
\describe{
\item{lower_limit}{lower limit of the nominal quantity range}
\item{upper_limit}{upper limit of the nominal quantity range}
\item{percentage_over_nominal_quantity}{percentage of error over nominal quantity (if available)}
\item{absolute_value}{absolute value of the maximum error tolerance (if available)}
}
}
\usage{
max_errors
}
\description{
A dataset containing the maximum error tolerances for different nominal
quantities, according to the specifications provided in the table.
}
\keyword{datasets}
