% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{mean_analysis}
\alias{mean_analysis}
\title{Perform Mean Analysis on Sample Data}
\usage{
mean_analysis(df, col_name, batch_size, nominal_quantity)
}
\arguments{
\item{df}{A data frame containing the sample data for analysis.}

\item{col_name}{A string representing the name of the column to analyze within the data frame.}

\item{batch_size}{A numeric value indicating the size of the batch being produced.}

\item{nominal_quantity}{A numeric value indicating the nominal quantity of the product being produced.}
}
\value{
A string indicating whether the batch should be accepted or rejected based on the mean analysis.
}
\description{
Given a data frame of sample data, a column name, a nominal quantity, and a batch size,
calculates the mean of the column and determines whether the batch should be
accepted or rejected based on the comparison of the mean with the nominal quantity minus
the estimation of the standard deviation.
}
\examples{
df <- data.frame(values = rep(10,30))
result <- mean_analysis(df, "values", 500, 10)
result

}
