% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_analysis_params}
\alias{mean_analysis_params}
\title{Mean analysis parameters}
\format{
A data frame with 2 rows and 4 variables:
\describe{
\item{batch_size_min}{lower limit of the batch size interval}
\item{batch_size_max}{upper limit of the batch size interval}
\item{sample_sizes}{sample sizes required for the mean analysis for each batch size}
\item{computed_student_distribution}{Student distribution computed for the corresponding sample sizes}
}
}
\usage{
mean_analysis_params
}
\description{
A dataset containing the required sample sizes for mean analysis and the corresponding computed Student distribution
using the specification on RD 1802/20XX.
}
\keyword{datasets}
