% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noncon_analysis_limits}
\alias{noncon_analysis_limits}
\title{Acceptance and Rejection limits by batch size}
\format{
A data frame with 3 rows and 6 variables:
\describe{
\item{batch_size_min}{The lower limit of the batch size interval.}
\item{batch_size_max}{The upper limit of the batch size interval.}
\item{first_analysis_acceptance}{The number of pieces that pass the first analysis and are accepted.}
\item{first_analysis_rejection}{The number of pieces that fail the first analysis and are rejected.}
\item{second_analysis_acceptance}{The number of pieces that pass the second analysis and are accepted.}
\item{second_analysis_rejection}{The number of pieces that fail the second analysis and are rejected.}
}
}
\usage{
noncon_analysis_limits
}
\description{
A dataset containing the acceptance and rejection limits for batches of different sizes
according to the specifications of RD 1802/20XX.
}
\keyword{datasets}
