% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{noncon_limit_value}
\alias{noncon_limit_value}
\title{Limit Acceptance Value}
\usage{
noncon_limit_value(nominal_quantity)
}
\arguments{
\item{nominal_quantity}{A numeric value indicating the nominal quantity for which
the maximum acceptable value is to be calculated.}
}
\value{
A numeric value indicating the maximum acceptable value for the given
  nominal quantity, which is twice the maximum error tolerance calculated by the
  "noncon_thres_value" function. If the nominal quantity is outside the range of the
  reference table, the function returns an error message.
}
\description{
Calculates the maximum acceptable value for a given nominal quantity, based on a
reference table of maximum errors.
}
\examples{
noncon_limit_value(20)

noncon_limit_value(5000)
}
