% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{noncon_thres_value}
\alias{noncon_thres_value}
\title{Nonconformity Threshold Value}
\usage{
noncon_thres_value(nominal_quantity)
}
\arguments{
\item{nominal_quantity}{A numeric value indicating the nominal quantity for which
the maximum error tolerance is to be calculated.}
}
\value{
A numeric value indicating the maximum error tolerance for the given
nominal quantity. If the nominal quantity is outside the range of the reference
table, the function returns an error message.
}
\description{
Calculates the maximum error tolerance for a given nominal quantity, based on a
reference table of maximum errors.
}
\examples{
noncon_thres_value(20)

noncon_thres_value(5000)
}
