% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_dataset}
\alias{sample_dataset}
\title{Sample dataset}
\format{
A data frame with 2 rows and 4 variables:
\describe{
\item{Id}{Id of the package}
\item{Paquete}{Package}
\item{Volumen}{Effective quantity of the package}
\item{Fecha}{Date of the measurement}
}
}
\usage{
sample_dataset
}
\description{
A dataset containing an example to run the package.
}
\keyword{datasets}
