% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_sizes}
\alias{sample_sizes}
\title{Sample sizes corresponding to each batch size.}
\format{
A data frame with 3 rows and 4 variables:
\describe{
  \item{batch_size_min}{lower limit of the batch size interval}
  \item{batch_size_max}{upper limit of the batch size interval}
  \item{first_analysis_sample }{sample needed for the first non-conformities check}
  \item{second_analysis_sample }{sample needed for the second non-conformities check}
  \item{mean_analysis_sample }{sample needed for the mean analysis}
}
}
\usage{
sample_sizes
}
\description{
A dataset containing the sample sizes corresponding to each batch size
 using the specification on RD 1802/20XX
}
\keyword{datasets}
