% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_core.R
\name{std_dev_estimate}
\alias{std_dev_estimate}
\title{Calculate the standard deviation estimate of a column in a data frame}
\usage{
std_dev_estimate(df, col_name)
}
\arguments{
\item{df}{A data frame containing the column to calculate the standard deviation estimate.}

\item{col_name}{A string representing the name of the column for which to calculate the estimate.}
}
\value{
A numeric value representing the estimate of the standard deviation of the column.
}
\description{
Given a data frame and the name of a column, calculates the estimate of the standard deviation
using the following formulas:
1. Estimate of variance = (corrected sum of squares)/(n - 1)
2. Corrected sum of squares = sum(x^2) - (sum(x)^2)/n
3. Estimate of standard deviation = sqrt(estimate of variance)
}
\examples{
data(mtcars)
std_dev_estimate <- std_dev_estimate(mtcars, "mpg")
std_dev_estimate
}
